smartcare_backend/requirements.txt
flask==2.1.2
flask-mongoengine==1.0.0
python-dotenv==0.20.0
WTForms==3.0.1
PyJWT==1.7.1
Flask-Bcrypt==1.0.1
